Move to avoid_use
